GraphicalModels

GraphicalModels.FGNode

GraphicalModels.FactorGraph

GraphicalModels.FactorGraph

GraphicalModels.FactorNode

GraphicalModels.Variable

GraphicalModels.VariableNode

Base.length

GraphicalModels.FGNode — Type

A node of a Factor Graph.

GraphicalModels.FactorGraph — Type

A Factor Graph is a bipartite graph where nodes are either variables or factors.

Arguments

variables: vector of variable nodes.

factors: vector of factor nodes.

GraphicalModels.FactorGraph — Method

FactorGraph(filename::AbstractString)::FactorGraph
FactorGraph(io::IO=stdin)::FactorGraph

Reads a model from file in UAI Competition Format and returns the correspoding factor graph. See https://www.cs.huji.ac.il/project/UAI10/fileFormat.php for details about the file format.

GraphicalModels.FactorNode — Type

Representes a factor node.

Arguments

factor: a multidimensional array representing a function of the neighbors. Each dimension corresponds to the order of the variable in the vector neighbors

neighbors: adjacent variable nodes

GraphicalModels.Variable — Type

A Discrete Variable.

GraphicalModels.VariableNode — Type

Represents a variable node.

Arguments

variable: integer identifying variable

neighbors: adjacent factor nodes

Base.length — Method

Returns the number of values of the variable.